requirements.txt
pip>=10.0.1
python-telegram-bot>=11.1.0
IMDbPY >= 6.6

